D3D9: ROOT CAUSE of the speckle — GPU-written guest memory never reaches CPU RAM
NX1 moves texture content through the GPU: memexport writes and render-to-texture
resolves. On hardware those land in unified memory and the CPU sees them. Here
they land in the reference backend's GPU-side shared memory buffer, and the copy
back to CPU-visible guest RAM is a readback feature that is OFF by default
(readback_memexport=false, readback_resolve="none"). The reference marks the
ranges RangeWrittenByGpu and samples its own GPU buffer -- which is why it always
rendered clean. The game's CPU-side bakes (and our decoder) read guest RAM, which
still holds the previous occupant's bytes: structured garbage, byte-stable,
present from the first decode, invisible to the page-protection write-watch.

Proof, in-game: readback_memexport=true alone visibly fixed the wall class;
adding readback_resolve="full" + nx1_skip_reference_raster=false (the reference
must actually rasterize or its EDRAM is junk) made the scene nearly completely
clean -- until ~16 GB of process memory, when the readback path begins failing
SILENTLY (RequestReadbackBuffer null -> skip) and speckle returned. Coherency
lasted exactly as long as allocations did.

Instrument fixes that got us there:
  - DECODECHANGE/DECODEHIST keyed by VIEW (DecodeViewKey: address + every
    decode-affecting fetch field). Address-alone keying fabricated 1359 phantom
    "content changed with zero writes" reports from multi-view binds of one
    address; the fixed instrument cut that to a real trickle -- which was the
    readback landing content.
  - CPFETCH/D9FETCH: draw-sequence fetch logs on each side's own clock (the CP
    at its draw from the register file, us at ours from the shadow), aligned
    offline. Replaces the invalid instantaneous register-file comparison, whose
    "mismatches" were echoes of our own earlier draws (ring lag).

Playable config for now: readback_memexport=true (memexport draws execute in the
reference CP even with raster skipped; SM3 cannot run them natively, so this
stays), readback_resolve=none + skip_reference_raster=true (resolve readback
must stay off while raster is skipped -- it would copy never-drawn EDRAM junk).
The resolve-baked class (dumpster lids, roof imposters) remains speckled in this
config. Next work item: native resolve->guest-RAM writeback in the D3D9 renderer,
which already executes every resolve and knows the destination address, pitch and
format -- async staged readback, deduped per destination, so the reference can
stay fully skipped with no memory or perf cliff.

+++ nx1_mp/src/d3d9/d3d9_resources.h
@@ -509,9 +509,11 @@ class ResourceTracker {
   /// READ that memory faithfully. This answers the other question directly: was the prefix ever
   /// written at all? Pages with zero writes were never touched by the guest while we watched.
   std::vector<uint8_t> page_writes_;
-  /// Last decoded content hash per texture address, so a texture that decodes one way and then
-  /// differently later reports the transition (see DECODECHANGE). Keyed on address alone --
-  /// this is a diagnostic, and the pool reassigning an address is itself worth seeing.
+  /// Last decoded content hash per texture VIEW (address + every fetch-constant field that
+  /// changes the decoded bytes -- see DecodeViewKey), so a texture that decodes one way and
+  /// then differently later reports the transition (see DECODECHANGE). Address-alone keying
+  /// was a bug that reported phantom changes whenever the same memory was bound through two
+  /// views, which reads exactly like writes invisible to page protection.
   struct DecodeStamp {
     uint64_t hash = 0;
     uint64_t frame = 0;
@@ -525,7 +527,7 @@ class ResourceTracker {
     uint32_t decodes = 0;
     uint32_t changes = 0;
   };
-  std::unordered_map<uint32_t, DecodeStamp> decode_hashes_;
+  std::unordered_map<uint64_t, DecodeStamp> decode_hashes_;
   /// Packed-mip fix instrumentation: decodes of <=16 texel textures, decodes whose fetch constant
   /// declares a packed mip tail, and decodes that actually got a non-zero sub-tile offset.
   uint64_t small_decodes_ = 0;
